split datatypes definitions out from router...
...in preparation to ghcjs client side library, which we do
not want to burden with inappropriate dependencies such as
wai.
